Move some common functionality to interface
