+++ stripecardscan/src/main/java/com/stripe/android/stripecardscan/cardimageverification/CardImageVerificationActivity.kt
@@ -4,7 +4,6 @@ import android.annotation.SuppressLint
 import android.content.Intent
 import android.graphics.Typeface
 import android.os.Bundle
-import android.util.Log
 import android.util.Size
 import android.view.Gravity
 import android.view.View
@@ -13,7 +12,6 @@ import android.widget.TextView
 import androidx.annotation.Keep
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.constraintlayout.widget.ConstraintSet
-import com.stripe.android.camera.framework.AnalyzerLoopErrorListener
 import com.stripe.android.camera.framework.Stats
 import com.stripe.android.stripecardscan.R
 import com.stripe.android.stripecardscan.cardimageverification.exception.InvalidCivException
@@ -37,14 +35,18 @@ import com.stripe.android.stripecardscan.payment.card.getIssuerByDisplayName
 import com.stripe.android.stripecardscan.payment.card.isValidPanLastFour
 import com.stripe.android.stripecardscan.payment.card.lastFour
 import com.stripe.android.stripecardscan.scanui.CancellationReason
+import com.stripe.android.stripecardscan.scanui.ScanErrorListener
 import com.stripe.android.stripecardscan.scanui.ScanResultListener
 import com.stripe.android.stripecardscan.scanui.SimpleScanActivity
+import com.stripe.android.stripecardscan.scanui.SimpleScanStateful
+import com.stripe.android.stripecardscan.scanui.SimpleScanStateful.ScanState
 import com.stripe.android.stripecardscan.scanui.util.getColorByRes
 import com.stripe.android.stripecardscan.scanui.util.getDrawableByRes
 import com.stripe.android.stripecardscan.scanui.util.hide
 import com.stripe.android.stripecardscan.scanui.util.setTextSizeByRes
 import com.stripe.android.stripecardscan.scanui.util.setVisible
 import com.stripe.android.stripecardscan.scanui.util.show
+import com.stripe.android.stripecardscan.scanui.util.startAnimation
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
@@ -74,7 +76,14 @@ internal data class RequiredCardDetails(
 private val MINIMUM_RESOLUTION = Size(1067, 600) // minimum size of OCR
 
 @Keep
-internal open class CardImageVerificationActivity : SimpleScanActivity<RequiredCardDetails?>() {
+internal open class CardImageVerificationActivity :
+    SimpleScanActivity<RequiredCardDetails?>(), SimpleScanStateful {
+
+    override var scanState: ScanState = ScanState.NotFound
+
+    override var scanStatePrevious: ScanState? = null
+
+    override val scanErrorListener: ScanErrorListener = ScanErrorListener()
 
     /**
      * The text view that lets a user indicate they do not have possession of the required card.
@@ -271,9 +280,16 @@ internal open class CardImageVerificationActivity : SimpleScanActivity<RequiredC
         }
 
         cannotScanTextView.setOnClickListener { userCannotScan() }
+
+        displayState(scanState, scanStatePrevious)
     }
 
-    private fun ensureValidParams() = when {
+    override fun onResume() {
+        super.onResume()
+        scanState = ScanState.NotFound
+    }
+
+    override fun ensureValidParams() = when {
         params.stripePublishableKey.isEmpty() -> {
             scanFailure(InvalidStripePublishableKeyException("Missing publishable key"))
             false
@@ -499,7 +515,52 @@ internal open class CardImageVerificationActivity : SimpleScanActivity<RequiredC
     }
 
     override fun displayState(newState: ScanState, previousState: ScanState?) {
-        super.displayState(newState, previousState)
+        when (newState) {
+            is ScanState.NotFound -> {
+                viewFinderBackgroundView
+                    .setBackgroundColor(getColorByRes(R.color.stripeNotFoundBackground))
+                viewFinderWindowView
+                    .setBackgroundResource(R.drawable.stripe_card_background_not_found)
+                viewFinderBorderView.startAnimation(R.drawable.stripe_card_border_not_found)
+                instructionsTextView.setText(R.string.stripe_card_scan_instructions)
+                cardNumberTextView.hide()
+                cardNameTextView.hide()
+            }
+            is ScanState.FoundShort -> {
+                viewFinderBackgroundView
+                    .setBackgroundColor(getColorByRes(R.color.stripeFoundBackground))
+                viewFinderWindowView
+                    .setBackgroundResource(R.drawable.stripe_card_background_found)
+                viewFinderBorderView.startAnimation(R.drawable.stripe_card_border_found)
+                instructionsTextView.setText(R.string.stripe_card_scan_instructions)
+                instructionsTextView.show()
+            }
+            is ScanState.FoundLong -> {
+                viewFinderBackgroundView
+                    .setBackgroundColor(getColorByRes(R.color.stripeFoundBackground))
+                viewFinderWindowView
+                    .setBackgroundResource(R.drawable.stripe_card_background_found)
+                viewFinderBorderView.startAnimation(R.drawable.stripe_card_border_found_long)
+                instructionsTextView.setText(R.string.stripe_card_scan_instructions)
+                instructionsTextView.show()
+            }
+            is ScanState.Correct -> {
+                viewFinderBackgroundView
+                    .setBackgroundColor(getColorByRes(R.color.stripeCorrectBackground))
+                viewFinderWindowView
+                    .setBackgroundResource(R.drawable.stripe_card_background_correct)
+                viewFinderBorderView.startAnimation(R.drawable.stripe_card_border_correct)
+                instructionsTextView.hide()
+            }
+            is ScanState.Wrong -> {
+                viewFinderBackgroundView
+                    .setBackgroundColor(getColorByRes(R.color.stripeWrongBackground))
+                viewFinderWindowView
+                    .setBackgroundResource(R.drawable.stripe_card_background_wrong)
+                viewFinderBorderView.startAnimation(R.drawable.stripe_card_border_wrong)
+                instructionsTextView.setText(R.string.stripe_scanned_wrong_card)
+            }
+        }
 
         when (newState) {
             is ScanState.NotFound, ScanState.FoundShort, ScanState.FoundLong, ScanState.Wrong -> {
@@ -527,17 +588,5 @@ internal open class CardImageVerificationActivity : SimpleScanActivity<RequiredC
             scanStatistics = ScanStatistics.fromStats()
         )
         super.closeScanner()
-    }
-
-    private val scanErrorListener = object : AnalyzerLoopErrorListener {
-        override fun onAnalyzerFailure(t: Throwable): Boolean {
-            Log.e(Config.logTag, "Error executing analyzer", t)
-            return false
-        }
-
-        override fun onResultFailure(t: Throwable): Boolean {
-            Log.e(Config.logTag, "Error executing result", t)
-            return true
-        }
     }
 }
